+++ src/cobalt/types/utils.rs
@@ -422,7 +422,7 @@ pub fn bin_op<'ctx>(mut lhs: Variable<'ctx>, mut rhs: Variable<'ctx>, op: &str, 
                         (Some(PointerValue(l)), Some(IntValue(r)), SizeType::Static(x), false) => {
                             let pt = ctx.context.i64_type();
                             let v1 = ctx.builder.build_load(l, "").into_pointer_value();
-                            let v2 = ctx.builder.build_int_mul(r, pt.const_int(x, false), "");
+                            let v2 = ctx.builder.build_int_mul(r, pt.const_int(x as u64, false), "");
                             let v3 = ctx.builder.build_ptr_to_int(v1, pt, "");
                             let v4 = ctx.builder.build_int_add(v3, v2, "");
                             let v5 = ctx.builder.build_int_to_ptr(v4, b.llvm_type(ctx).unwrap().ptr_type(inkwell::AddressSpace::from(0u16)), "");
@@ -439,7 +439,7 @@ pub fn bin_op<'ctx>(mut lhs: Variable<'ctx>, mut rhs: Variable<'ctx>, op: &str, 
                         (Some(PointerValue(l)), Some(IntValue(r)), SizeType::Static(x), false) => {
                             let pt = ctx.context.i64_type();
                             let v1 = ctx.builder.build_load(l, "").into_pointer_value();
-                            let v2 = ctx.builder.build_int_mul(r, pt.const_int(x, false), "");
+                            let v2 = ctx.builder.build_int_mul(r, pt.const_int(x as u64, false), "");
                             let v3 = ctx.builder.build_ptr_to_int(v1, pt, "");
                             let v4 = ctx.builder.build_int_sub(v3, v2, "");
                             let v5 = ctx.builder.build_int_to_ptr(v4, b.llvm_type(ctx).unwrap().ptr_type(inkwell::AddressSpace::from(0u16)), "");
@@ -895,7 +895,7 @@ pub fn bin_op<'ctx>(mut lhs: Variable<'ctx>, mut rhs: Variable<'ctx>, op: &str, 
                     (Some(PointerValue(l)), Some(IntValue(r)), SizeType::Static(x), false) => Some({
                         let pt = ctx.context.i64_type();
                         let v1 = ctx.builder.build_ptr_to_int(l, pt, "");
-                        let v2 = ctx.builder.build_int_mul(r, pt.const_int(x, false), "");
+                        let v2 = ctx.builder.build_int_mul(r, pt.const_int(x as u64, false), "");
                         let v3 = ctx.builder.build_int_add(v1, v2, "");
                         PointerValue(ctx.builder.build_int_to_ptr(v3, b.llvm_type(ctx).unwrap().ptr_type(inkwell::AddressSpace::from(0u16)), ""))
                     }),
@@ -910,7 +910,7 @@ pub fn bin_op<'ctx>(mut lhs: Variable<'ctx>, mut rhs: Variable<'ctx>, op: &str, 
                     (Some(PointerValue(l)), Some(IntValue(r)), SizeType::Static(x), false) => Some({
                         let pt = ctx.context.i64_type();
                         let v1 = ctx.builder.build_ptr_to_int(l, pt, "");
-                        let v2 = ctx.builder.build_int_mul(r, pt.const_int(x, false), "");
+                        let v2 = ctx.builder.build_int_mul(r, pt.const_int(x as u64, false), "");
                         let v3 = ctx.builder.build_int_sub(v1, v2, "");
                         PointerValue(ctx.builder.build_int_to_ptr(v3, b.llvm_type(ctx).unwrap().ptr_type(inkwell::AddressSpace::from(0u16)), ""))
                     }),
@@ -928,7 +928,7 @@ pub fn bin_op<'ctx>(mut lhs: Variable<'ctx>, mut rhs: Variable<'ctx>, op: &str, 
                     (Some(PointerValue(l)), Some(IntValue(r)), SizeType::Static(x), false) => Some({
                         let pt = ctx.context.i64_type();
                         let v1 = ctx.builder.build_ptr_to_int(l, pt, "");
-                        let v2 = ctx.builder.build_int_mul(r, pt.const_int(x, false), "");
+                        let v2 = ctx.builder.build_int_mul(r, pt.const_int(x as u64, false), "");
                         let v3 = ctx.builder.build_int_add(v1, v2, "");
                         PointerValue(ctx.builder.build_int_to_ptr(v3, b.llvm_type(ctx).unwrap().ptr_type(inkwell::AddressSpace::from(0u16)), ""))
                     }),
@@ -1333,7 +1333,7 @@ pub fn pre_op<'ctx>(mut val: Variable<'ctx>, op: &str, ctx: &CompCtx<'ctx>) -> O
                             let pt = ctx.context.i64_type();
                             let v1 = ctx.builder.build_load(v, "").into_pointer_value();
                             let v2 = ctx.builder.build_ptr_to_int(v1, pt, "");
-                            let v3 = ctx.builder.build_int_add(v2, pt.const_int(x, false), "");
+                            let v3 = ctx.builder.build_int_add(v2, pt.const_int(x as u64, false), "");
                             let v4 = ctx.builder.build_int_to_ptr(v3, b.llvm_type(ctx).unwrap().ptr_type(inkwell::AddressSpace::from(0u16)), "");
                             ctx.builder.build_store(v, v4);
                         }
@@ -1346,7 +1346,7 @@ pub fn pre_op<'ctx>(mut val: Variable<'ctx>, op: &str, ctx: &CompCtx<'ctx>) -> O
                             let pt = ctx.context.i64_type();
                             let v1 = ctx.builder.build_load(v, "").into_pointer_value();
                             let v2 = ctx.builder.build_ptr_to_int(v1, pt, "");
-                            let v3 = ctx.builder.build_int_sub(v2, pt.const_int(x, false), "");
+                            let v3 = ctx.builder.build_int_sub(v2, pt.const_int(x as u64, false), "");
                             let v4 = ctx.builder.build_int_to_ptr(v3, b.llvm_type(ctx).unwrap().ptr_type(inkwell::AddressSpace::from(0u16)), "");
                             ctx.builder.build_store(v, v4);
                         }
